requirements.txt
requests
streamlit
matplotlib
python-dotenv
